feat(git): default-branch alias
Summary:

A new alis was added, `git default-branch` which will return master or
main depending on which appears first.

Impact:

Many repositories are moving over to using a default branch of "main"
from the archaic "master" (for very good reasons). However, not all
repositories are there yet. In order for dotfiles to work across
different repositories, it's useful to abstract the default branch
behind an alias so that other commands work with either/or.